Minor fix in get_avd_name_for_emulator_port

+++ devices/avd_manager.py
@@ -12,7 +12,7 @@ class AvdManager(object):
         self.avdmanager_path = RequiredFeature('avdmanager_path').request()
 
     def get_avd_name_for_emulator_port(self, port):
-        avd_index = (port - 5554) / 2
+        avd_index = int((port - 5554) / 2)
         return self.avd_series + "_" + str(avd_index)
 
     def avd_name_exists(self, avd_name):
